refac: Refactor Leaderboard Components and Optimize Rank Rewards Rendering
